-Use multi-catch -Unnecessary unboxing

+++ java/j2ee.persistence/src/org/netbeans/modules/j2ee/persistence/entitygenerator/SQLType.java
@@ -135,7 +135,7 @@ class SQLType {
     private Class getClassForCharType (Integer length, boolean isNullable) {
         switch (sqlType) {
             case Types.CHAR:
-                if ((length != null) && (length.intValue() == 1)) {
+                if ((length != null) && (length == 1)) {
                      return typeList[1];//typeList[isNullable ? 1 : 2] is replaced with String always because of issue #195674 (openjpa fo Character handling)
                 }
             default:
@@ -148,8 +148,8 @@ class SQLType {
     // the no-arg getMemberType method are sufficient
     private Class getClassForNumericType(Integer precision, 
             Integer scale, boolean isNullable) {
-        int precValue = ((precision == null) ? -1 : precision.intValue());
-        int scaleValue = ((scale == null) ? -1 : scale.intValue());
+        int precValue = ((precision == null) ? -1 : precision);
+        int scaleValue = ((scale == null) ? -1 : scale);
 
         switch (sqlType) {
             case Types.DECIMAL:
